refactor(palace): split palace.py into a package (#2494)
* refactor(palace): split palace.py into a package

The 1800-line shared palace module is now mempalace/palace/ so PRs can
target collection access, closets, mine locks, or mined-set checks
without colliding.

The public import path is unchanged: fragments exec into the package
namespace so get_collection, mine_lock, mine_palace_lock, and test
monkeypatches on process globals keep working.

* docs(palace): address Copilot nits on the package split

Fix article in CHANGELOG, describe the package in the docstring, point CLOSETS.md and RFC 002 at mempalace.palace, and correct a moved docstring typo.
