+++ tests/test_readme_claims.py
@@ -46,6 +46,12 @@ def _searcher_source() -> str:
     return "\n".join(_read(path) for path in sorted(pkg.glob("*.py")))
 
 
+def _palace_source() -> str:
+    """Concatenate palace package sources."""
+    pkg = MEMPALACE_PKG / "palace"
+    return "\n".join(_read(path) for path in sorted(pkg.glob("*.py")))
+
+
 def _tools_dict_keys() -> list:
     """Return the list of tool names registered in the TOOLS dict."""
     # Import the module-level TOOLS dict.  We can't just import mcp_server
@@ -154,7 +160,7 @@ class TestClosetsExist:
     def test_get_closets_collection_exists(self):
         """Claim: closets are a shipped feature.
         palace.py must export get_closets_collection()."""
-        src = _read(MEMPALACE_PKG / "palace.py")
+        src = _palace_source()
         assert "def get_closets_collection(" in src, (
             "palace.py does not define get_closets_collection(). "
             "Closets are described in README but the collection function is missing."
@@ -234,7 +240,7 @@ class TestEntityMetadataExtraction:
         """Claim: entity extraction is part of the mining pipeline.
         Either miner.py or palace.py must extract entities."""
         miner_src = _read(MEMPALACE_PKG / "miner.py")
-        palace_src = _read(MEMPALACE_PKG / "palace.py")
+        palace_src = _palace_source()
         # Entity extraction can be in either file — palace.py has it for closets
         has_entity_extraction = (
             "entities" in palace_src and "_ENTITY_STOPLIST" in palace_src
@@ -398,7 +404,7 @@ class TestMineLock:
     def test_mine_lock_exists(self):
         """Claim: multi-agent file locking is shipped.
         palace.py must define mine_lock."""
-        src = _read(MEMPALACE_PKG / "palace.py")
+        src = _palace_source()
         assert "def mine_lock(" in src, (
             "palace.py does not define mine_lock(). "
             "Multi-agent locking is claimed as shipped but function is missing."
@@ -412,7 +418,7 @@ class TestMineLock:
 
     def test_mine_lock_is_context_manager(self):
         """mine_lock should be a context manager (used with `with` statement)."""
-        src = _read(MEMPALACE_PKG / "palace.py")
+        src = _palace_source()
         # It should be decorated with @contextlib.contextmanager or similar
         # Find the mine_lock definition and check for context manager pattern
         assert "@contextlib.contextmanager" in src or "def __enter__" in src, (
